Volume backup response schemas are incorrect
Change I2ab6ecfeb0c48c7ef4a0fc94e670d43dbc3f6650 added response
schemas for volume backup responses. The schema specifies the 'name'
as type string in the response for create-backup and update-backup,
but the name is not required in the create request and the cinder
backup service does not supply one. This element should be string
or null.

Change-Id: I032e54a2065859d4ebc0969faccf68494fe745bc
Closes-bug: #1894844