New list config option to list network features without API extension
Historically Neutron had API extensions to make it clear what feature
is available and what is not. There are however features without new
API extensions, see [1] and [2].
To avoid having lots of config options that makes it possible in a
branchless tempest to detect if a given feature is available or not on
a given Neutron branch this patch proposes a list config option similar
to api_extensions: available_features.
available_features is a list option with the special all entry, and new
tests that depend on API extensionless features can use this to be
skipped on older branches.

[1] https://review.opendev.org/750355
[2] https://review.opendev.org/743695

Change-Id: I9d4be68db0f18dcaded4bfe78f5a314cb450987a