Also calculate external subnets when check cidr whether in use
The external network be shared by all tenant. When it associated
to a tenant router the cidr of it's subnets cannot overlap with
tenant subnets', otherwise it will lead to bug #1903433. The related
fix patch: https://review.opendev.org/761829

So, in tempest, the external subnets need was considered when check
cidr whether in use.

Related-Bug: #1903433
Change-Id: Ief84251adb0368533570d9e9f6630de92e27c5c3