compute: Cleanup server before image in AttachSCSIVolumeTestJSON
This change simply reorders the cleanup operations within the
test_attach_scsi_disk_with_config_drive test to ensure the server is
always removed prior to the image.

This is required as when both Glance and Nova use a shared RBD backend
Nova will clone an instance disk directly from the Glance RBD volume.
This results in any attempt to delete the Glance image to fail while the
server is still provisioned as it still references the Glance RBD volume.

Change-Id: Ibcb1d1e3d1bb087060bcd0569a70c9955cd1357f
Closes-Bug: #1905725