Use victoria stable constraint in tox to release 26.0.0
We are going to release Tempest new tag 26.0.0 to declare
the end of support for stable/stein. So that new tag use
stable/victoria constraint in the tox env as victoria constraints
for sure will work with tag. It did not use wallaby constraints
as wallaby dev cycle is in progress and so does their constraints.

Further information:
 https://docs.openstack.org/tempest/latest/requirement_upper_constraint_for_tempest.html

Change-Id: I31433c15984e38e5651b2f3d10c8769f59e30cc1