Fix race condition for image test
If first image is not captured in saving state,
then we should skip the test
- test_create_second_image_when_first_image_is_being_saved

Test is for creating the another image is first one is in
saving state and if image creation operation is fast and it
gets created before second API request then this test will
fail on its assert.

To handle the test scope and this race condition, we need to
assert on second operation if  image is in saving state.This
can be done via capturing the timeout exception from first image
creation and waiting for saving state.

Change-Id: Id59f7ebb223f968109917182e29267faa808a072
Closes-Bug: #1881592