Test glance distributed import
Glance is developing functionality to solve the problem where an
image is staged against one worker and the import call arrives at
the other. This adds support to test that. The dependent devstack
patch sets up another glance worker in the catalog as "image_remote",
which we can use to ensure that stage and import calls arrive at
different workers.

Change-Id: I38bafb5316ce3b8aac9b8ac11ae27110a93d3ebd
Depends-On: https://review.opendev.org/c/openstack/devstack/+/770487