Fix memory explosion in multi-store image tests
For some reason, the MultiStoresImportImages test was causing the
test runner to balloon to ~1.5G, which generates spurious OOMs in
the gate, resulting in worker death and failure.

This test was generating 10M of data with data_utils.random_bytes(),
instead of the default of 1K, for no apparent reason. We are still
not sure why, but this results in the memory ballooning, potentially
because of BytesIO, or request.put() mishandling.

Regardless, this is unnecessary, and causing failures in the gate,
so this patch switches it back to the default, which works fine.

Change-Id: I9f1ea0114531a735bd38ad54da7ce15fb2bf7f7c