test_encrypted_cinder_volumes: Do not create an image for this test
As with I50d3b874b7e0676436c434be6c059f221041560f for the
cinder-tempest-plugin this encrypted volume test does not require a
unique image to create an instance. The creation leading to test
failures in the nova-ceph-multistore job that sets
[workarounds]/never_download_image_if_on_rbd while valdiating some edge
usecases.

This change simply removes the creation of the image and relys
instead on the behaviour of tempest.common.compute.create_test_server
that uses CONF.compute.image_ref when an image_id is not provided [1].

[1] https://opendev.org/openstack/tempest/src/branch/master/tempest/common/compute.py#L108-L109

Change-Id: Iee0d863770def63df8601222265dc22226457c0c
Related-Bug: #1912607