Fix system & domain scoped admin dynamic credential
With I8bebb5b9b6d8da62e6a5268d827787da461cc0d6 Tempest
started supporting the system and domain scope along
with project. But system and domain admin are not created
with requried scope. If admin role is requested even with
system or domain scope get_credentials() method does not pass
the requested scope to _create_creds() so it is always created
with project scope.

This commit fix this by passing the correct scope while creating
the cred.

Change-Id: Id2c9f5b304106a5be15639a69f95be424a394436