Add config option for enforce_scope
As servies are moving towards new RBAC defaults
and Tempest or plugins are writting the new
tests or moving the existing tests towards new defaults
policy, we need to add config option to enbale the
scoped token.

Adding it in Tempest configwill help to avoid any conflict
in same config name. This new config option will be used
in Tempest tests or in Tempest plugins side too.

Change-Id: I788e25c499410f452bcfaeab86611f597f2e50a7