Create default network for server advance scenario tests
Previously when fixing this bug for API tests,
we thought it cannot happen in scenario tests
with same reason as scenario base class do fetch
the specific network from neutron to pass it t nova
for creating the server.

But when 'CONF.network.port_vnic_type' and
'CONF.network.port_profile' are not configured in Tempest
then Tempest scenario manager does not fetch the network
from neutron.
- https://github.com/openstack/tempest/blob/fec2c93cdcc14ad08d0a35136ee287525e7a4879/tempest/scenario/manager.py#L242

In that case, tenant network is used, which is not present
as scenario manager does not create the default network
resource for credential and tenant.

- https://github.com/openstack/tempest/blob/fec2c93cdcc14ad08d0a35136ee287525e7a4879/tempest/scenario/manager.py#L298

This commit let server advance test (which is failing for multiple network
erorr) to create the default network which will be used to pass to the
nova API request while creating the test server.

Related-Bug: #1844568
Change-Id: I5a73a4f25f9092a4cdddcaf86f9ff2c89720409e