change ipv6 public network cidr on tempest-slow job
Neutron recently enhanced CIDR overlap checks for all the subnets
attached to router including external [1]. This breaks tempest
scenario ipv6 tests which is fixed in tempest by the following
commit [2]. However the fix mentioned in [2] cannot be ported
back to tempest 23.0.0 as rocky is in EM stage.

To fix this, the public IPv6 network CIDR is changed to use
different subnet on the tempest-slow job.

Closes-Bug: #1914037

[1] https://review.opendev.org/q/Id5d8ac09a38c656619f88a6f87b8f384fe4c55a8
[2] https://opendev.org/openstack/tempest/commit/4ad346e6497d4fbad9e379e8ed50730a0b6b55a3

Change-Id: I62a746cc871434e8488b19bc2ef5bbcfcf4f0be6