Enable devstack parallel installation in tempest-full-parallel job
Devstack depends-on patch enable the way to install the things
in parallel, let's try it in tempest-full-parallel job which
run all the scenario tests in parallel with API tests.

This can help to check the stability of devstack parallel logic
as well as to see how much improvement it add in runtime. If
tempest-full-parallel job run faster then it will be
bigger step to run back the scneario tests in parallel.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/771505
Change-Id: I1da50a89fb7617364249b04efcf207e7728f7de7