Add image task validation
Glance is adding a /image/$image/tasks API to make it easier to
consume the import API. This fetches the tasks after glance-direct
import, and validates that they look like we expect.

Because this is only supported after v2.12, we check for that before
doing the validation. To make that easier, this adds a has_version()
helper to the VersionsClient.

Change-Id: I2850f0659e82bf5c5a1005de0a063e7fcacadb51