Add option to support volume_type in volume creation
The volume creation uses default volume type from the backend. In
case of multiple backend, at the moment there is no option to test
volume test cases with volume type which is not default volume type.

Add a new config volume_type in tempest.conf.
Change the code to create volumes with volume_type mentioned in
the config.
Make sure when volume_type is not specified then volume will be
created with default volume type.

Change-Id: I1b4601e2ce6c83ec727b6dea65c67cacbfcfe957