Create tmd dir with specific prefix in check-uuid unit tests
test_fix_argument_yes for check-uuid tool try to
import the created tmp directory and work on that
which is failing with the below error:
ModuleNotFoundError: No module named 'tmpf61iim1k'

- https://88e1de9a81e55d590d5b-26f184bb59af339cfe698349cbda4177.ssl.cf5.rackcdn.com/770520/8/check/openstack-tox-cover/e3518e8/testr_results.html

It is happening more frequently nowadays: 40 occurrences in 7 days

http://logstash.openstack.org/#/dashboard/file/logstash.json

There is some race happening to delete the created tmp dir
which is then used to import the test file in check-uuid run
and raise error.

This commit try to create the tmp dir with specific prefix so that
any other tests cleanup deleting the *tmp* dir on root path then
it can avoid this failure.

Related-Bug: #1918316
Change-Id: Ibab610d3e59ec22bb8c37d66f262ed1d2648bbf8