Use stable constraint in tox to release new tag for Wallaby
We are going to release Tempest a new tag to declare the start of
support for stable/wallaby so the new tag should use stable/wallaby
constraint in the tox env.

For example: if anyone uses Tempest 27.0.0 in the future, say 1 year
later, then tox env also should use the wallaby contstraint instead
master for the compatibility.

These changes need to be moved back to master constraint once the
release is done. Using master constraint in tox.ini for all the older
tags is a problem and we faced the issue of constraint
incompatibility. Therefore, we have to work around it all over the
devstack, tempest, grenade and set the stable constraints via the
env var.

Change-Id: Id028b825519e9446e585503dddd39069c05c2867