Add wait to check the import task status
Once Glance finish the import operation(change image status
to active), it move the task to 'success' state but in between of
image become active and task is transitioning to 'success', tempest
try to check the task status and race condition happen.

Adding waiter method in test for task status check.

Closes-Bug: #1926671
Change-Id: I960b80314f1b0926eca33af830bc827f31cbeda6