Fix up the tempest-full-py3 job configurations
When we created the separate definition for tempest-full-py3 job
for swift enable on ussuri onwards, we missed to add the
few configutation like horizon, networking config due to rebase
and all
- I63159e5e8c0c8b6751ea481577b4c4637a7f25b5

This commit re-add all those configuration back to job version
of ussuri onwards.

This issue is found when we saw horizon dashbaord test is skipped
in tempest-full-py3 master gate.

- https://zuul.opendev.org/t/openstack/build/34892e0fd92f410bb99b37c8fe2e5255/log/job-output.txt#27738

Change-Id: I9602937f53933a9d30022512b7fe6a817d7e6472