zuul: Add a CentOS 8 stream job to integrated-gate-compute
As set out in the governance repo for Xena CentOS 8 stream with py36 is
a supported platform and runtime for the release:

https://governance.openstack.org/tc/reference/runtimes/xena.html

As a result the Nova team want to run CentOS 8 stream based jobs within
the integrated-gate-compute template.

An additional tempest-full-py3-centos-8-stream job is added to Tempest's
check and gate queues to ensure coverage here.

Both jobs are given additional swap to workaround bug #1949606, a
behaviour change in QEMU >= 5.0.0 when using [libvirt]virt_type=qemu
that causes additional memory to be consumed by each running instance.

This workaround of additional swap will be removed in the future once
Nova is able to workaround this itself through a new libvirt domain
configurable.

Change-Id: I4b074dca58960b2622dde15ad7d9a91b3ba0c096