Skip network creation for system and domain scoped token
For 'domain' and 'system' scoped token, there is no project_id
so we cannot create the network which need project id as one of
the parameter.

In Xena PTG[1], we have discussed about project mapped resource
creation/managing with system token
- L77-140 https://etherpad.opendev.org/p/nova-xena-ptg
- https://etherpad.opendev.org/p/consuming-system-scope

Once we sort out the network ceration or need for system, domain
scoped token then we can update the newtork creation in Tempest.

Change-Id: If6ae6465369c9018c716d48555fd99fc90ce0e59