Use wallaby stable constraint in tox to release 28.0.0
We are going to release Tempest new tag 28.0.0 to declare
the end of support for stable/train. So that new tag use
stable/wallaby constraint in the tox env as wallaby constraints
for sure will work with tag. It did not use xena constraints
as xena dev cycle is in progress and so does their constraints also
is in progress and can be bumped.

Further information:
 https://docs.openstack.org/tempest/latest/requirement_upper_constraint_for_tempest.html

Change-Id: I4abf58228eba0622f1122851ff398d49a200221e