Fix run-tempest-26 role for stable/stein
stable/stein uses old Tempest 26.0.0 to run the test
and that is why we use stable/stein upper-constraints
in the testing. But when any jobs with tempest plugins
try to run plugins tests, they fail.

This is because a new role run-tempest-26 (used for testing
with tempest 26.0.0) does not include the stable/stein in the
list of branches for whom we use stable constraints instead of
master constraints. and due to that run-tempest-26 finds constraints
mismatch and recreates the tempest virtual env. On recreation, venv
remove the previously installed tempest pluings and so does not run
their test.

Closes-Bug: #1938227
Change-Id: I00ac642fa602980d14612adc1ce65068d67db8ad