object_storage: retry object creation on Conflict
In some cases (such as when using Ceph RADOSGW with a load-balancer)
the container (bucket) creation will be received by different backend
to the object creation triggering a subtle edge case.

The rgw handling the object creation request has (cached) knowledge
the bucket doesn't exist (it didn't a few seconds ago) and will (in
rgw) create it, at which point triggering a "409 BucketAlreadyExists"
code.

Detecting this and retrying (with a small delay) should reduce
instances of this failing.

Change-Id: Id88fb93dd1a48d046917e9168d623e53497ec83e