hacking: Ignore H405
This patch proposes to ignore hacking H405, "multi line docstring
summary not separated with an empty line". This arbitrarily enforces
"title" lines for docstring that are more than 1 line, meaning a
simple doctring that needs just 2 lines will either need to be
condensed into a single line, or expanded to include a "title" and a
longer explanation paragraph underneath. This does not appear to bring
much value.

Change-Id: I1978c73ab6eda1dce85c96b612c4c5bff286efc6