Add --libpath argv option to fix unittest failure
Fixes intermittent tempest unittest failure of "test_fix_argument_yes".
The bug was in check_uuid.py and not in the test itself. As part of
the fix and backward compatibility, enabled check_uuid.py to run a
validation of a py module from any dir by adding --libpath parameter to
argv.

Closes-Bug:#1918316

Change-Id: If5d3a3852320e1d3b26af433f7a66be7d42e2b84