Remove compute api_extensions config option
Nova has remvoed the API extensions concept since
stein release
- https://blueprints.launchpad.net/nova/+spec/api-extensions-merge-stein

Nova cannot be configured to enable/disable extension now
and all extensions are enabled by default. Stein is not
supported by the Tempest master so it is safe to remove the
config optionn as well as the checks from tests.

Change-Id: I05c0a3b23d0824e1eb9a79353dcf2976c6bcf602