Set default value of --concurrency to 2
There is a non-negligible number of tests in tempest that need accounts
generated by account-generator with a concurrency of at least 2. Running
these tests with accounts generated with the current default value for
concurrency makes them therefore fail.

This change sets the default value for concurrency to 2 and updates
the documentation for --concurrency parameter.

Closes-Bug: #1948935
Change-Id: Iada980beff9db3c8cd02c84d0d7602264b66ddf8